enemy animation attack and damage

+++ FirstTopDownGame/Assets/Scripts/HealthBar.cs
@@ -8,19 +8,18 @@ public class HealthBar : MonoBehaviour
 
     Slider slider;
     
-
     void Awake()
 	{
         slider = GetComponent<Slider>();
     }
 
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
     {
         slider.maxValue = health;
         slider.value = health;
     }
 
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
